requirements.txt
flask==3.0.3
apscheduler==3.11.0
mysql-connector-python==9.0.0
psutil==5.9.8
pandas==1.5.3
openpyxl==3.1.5